// src/types.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export interface LevelConfig {
  worker: HTMLElement | null;
  lvlUpButton: HTMLElement | null;
  infoContainer: HTMLElement | null;
  addWorker: any;
  sellButton?: HTMLElement | null;
  container?: HTMLElement | null;
  converterButton?: HTMLElement | null;
}